requirements.txt
# Django Core
Django==4.2.16
Jinja2==3.1.4          
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.9

# Authentication & Tokens
djangorestframework-simplejwt==5.3.1

# Environment Variables
python-decouple==3.8

# Static Files
whitenoise==6.7.0

# Server
gunicorn==23.0.0

# Image Processing
Pillow==10.4.0

# Utilities
python-slugify==8.0.4

# Testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
coverage==7.6.4
